+++ lib/pages/journal/entry_details_page.dart
@@ -24,12 +24,10 @@ class EntryDetailPage extends StatefulWidget {
     super.key,
     @PathParam() required this.itemId,
     this.readOnly = false,
-    this.showBackIcon = true,
   });
 
   final String itemId;
   final bool readOnly;
-  final bool showBackIcon;
 
   @override
   State<EntryDetailPage> createState() => _EntryDetailPageState();