+++ lib/pages/settings/settings_page.dart
@@ -30,7 +30,10 @@ class _SettingsPageState extends State<SettingsPage> {
 
     return Scaffold(
       backgroundColor: colorConfig().bodyBgColor,
-      appBar: VersionAppBar(title: localizations.navTabTitleSettings),
+      appBar: VersionAppBar(
+        title: localizations.navTabTitleSettings,
+        showBackIcon: false,
+      ),
       body: Container(
         margin: const EdgeInsets.symmetric(
           vertical: 8,